requirements.txt
addict==2.4.0
albumentations==1.3.0
altair==4.2.0
appnope==0.1.3
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.3
decorator==5.1.1
easydict==1.10
entrypoints==0.4
executing==1.2.0
flatbuffers==22.10.26
fonttools==4.38.0
gitdb==4.0.9
GitPython==3.1.29
humanfriendly==10.0
idna==3.4
iglovikov-helper-functions==0.0.53
imagecorruptions==1.1.2
imageio==2.22.4
importlib-metadata==5.0.0
insightface==0.6.2
ipykernel==6.17.1
ipython==8.6.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jpeg4py==0.1.4
jsonschema==4.17.1
jupyter_client==7.4.7
jupyter_core==5.0.0
kiwisolver==1.4.4
kornia==0.6.8
lpips==0.1.4
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mpmath==1.2.1
nest-asyncio==1.5.6
networkx==2.8.8
numpy==1.23.5
onnx==1.12.0
onnxruntime==1.13.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
prettytable==3.5.0
prompt-toolkit==3.0.33
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pycparser==2.21
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qudida==0.0.4
requests==2.28.1
retinaface-pytorch==0.0.7
rich==12.6.0
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
stack-data==0.6.1
streamlit==1.15.1
sympy==1.11.1
threadpoolctl==3.1.0
tifffile==2022.10.10
toml==0.10.2
toolz==0.12.0
torch==1.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
wcwidth==0.2.5
zipp==3.10.0
